shadow: update to 4.1.0
With 4.1.0-1 the default behaviour of shadow is restored and a patch
from Gentoo [1] removed.
It may be debatable whether useradd should create a new group with the
same name as the username if the primary group is not specified with
-g, or not. For now we will follow the CRUX principle to use vanila
packages whenever possible.
However, there are plans to change something in that direction [2].

[1] http://bugs.gentoo.org/show_bug.cgi?id=128715
[2] http://lists.alioth.debian.org/pipermail/pkg-shadow-devel/2007-December/006252.html